Move the test helpers to NextLS.Support.Utils
